finalised code
